source-postgres: Move value translation into a different file
Long ago it made sense to combine discovery and value translation,
but I think we outgrew that point a while back.
