+++ source-postgres/discovery.go
@@ -2,26 +2,15 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"net"
 	"slices"
 	"sort"
 	"strings"
-	"time"
 
 	"github.com/estuary/connectors/sqlcapture"
 	"github.com/invopop/jsonschema"
 	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgtype"
 	"github.com/sirupsen/logrus"
-)
-
-const (
-	infinityTimestamp         = "9999-12-31T23:59:59Z"
-	negativeInfinityTimestamp = "0000-01-01T00:00:00Z"
-	rfc3339TimeFormat         = "15:04:05.999999999Z07:00"
-	truncateColumnThreshold   = 8 * 1024 * 1024 // Arbitrarily selected value
 )
 
 // DiscoverTables queries the database for information about tables available for capture.
@@ -253,210 +242,6 @@ func (db *postgresDatabase) TranslateDBToJSONType(column sqlcapture.ColumnInfo) 
 		jsonType.Description = *column.Description
 	}
 	return jsonType, nil
-}
-
-func translateRecordFields(table *sqlcapture.DiscoveryInfo, f map[string]interface{}) error {
-	if f == nil {
-		return nil
-	}
-	for id, val := range f {
-		var columnInfo *sqlcapture.ColumnInfo
-		if table != nil {
-			if info, ok := table.Columns[id]; ok {
-				columnInfo = &info
-			}
-		}
-
-		var translated, err = translateRecordField(columnInfo, val)
-		if err != nil {
-			return fmt.Errorf("error translating field %q value %v: %w", id, val, err)
-		}
-		f[id] = translated
-	}
-	return nil
-}
-
-func oversizePlaceholderJSON(orig []byte) json.RawMessage {
-	return json.RawMessage(fmt.Sprintf(`{"flow_truncated":true,"original_size":%d}`, len(orig)))
-}
-
-// translateRecordField "translates" a value from the PostgreSQL driver into
-// an appropriate JSON-encodeable output format. As a concrete example, the
-// PostgreSQL `cidr` type becomes a `*net.IPNet`, but the default JSON
-// marshalling of a `net.IPNet` isn't a great fit and we'd prefer to use
-// the `String()` method to get the usual "192.168.100.0/24" notation.
-func translateRecordField(column *sqlcapture.ColumnInfo, val interface{}) (interface{}, error) {
-	var columnName = "<unknown>"
-	var dataType interface{}
-	if column != nil {
-		columnName = column.Name
-		dataType = column.DataType
-	}
-
-	switch dataType {
-	case "timetz":
-		if x, ok := val.(string); ok {
-			var formats = []string{
-				"15:04:05.999999999Z07:00",
-				"15:04:05.999999999Z07",
-				"15:04:05Z07:00",
-				"15:04:05Z07",
-			}
-			for _, format := range formats {
-				if t, err := time.Parse(format, x); err == nil {
-					return t.Format(rfc3339TimeFormat), nil
-				}
-			}
-		}
-	}
-	switch x := val.(type) {
-	case net.HardwareAddr: // column types 'macaddr' and 'macaddr8'
-		return x.String(), nil
-	case [16]uint8: // column type 'uuid'
-		var s = new(strings.Builder)
-		for i := range x {
-			if i == 4 || i == 6 || i == 8 || i == 10 {
-				s.WriteString("-")
-			}
-			fmt.Fprintf(s, "%02x", x[i])
-		}
-		return s.String(), nil
-	case string:
-		if len(x) > truncateColumnThreshold {
-			return x[:truncateColumnThreshold], nil
-		}
-		return x, nil
-	case []byte:
-		if len(x) > truncateColumnThreshold {
-			return x[:truncateColumnThreshold], nil
-		}
-		return x, nil
-	case pgtype.Array[any]:
-		return translateArray(column, x)
-	case pgtype.Range[any]:
-		return stringifyRange(x)
-	case map[string]any: // TODO(wgd): Is this needed for anything after we fix JSON columns?
-		var bs, err = json.Marshal(x)
-		if err != nil {
-			return nil, fmt.Errorf("error reserializing json column %q: %w", columnName, err)
-		}
-		if len(bs) > truncateColumnThreshold {
-			return oversizePlaceholderJSON(bs), nil
-		}
-		return json.RawMessage(bs), nil
-	case pgtype.Text:
-		if len(x.String) > truncateColumnThreshold {
-			return x.String[:truncateColumnThreshold], nil
-		}
-		return x.String, nil
-	case pgtype.InfinityModifier:
-		// Postgres has special timestamp values for "infinity" and "-infinity" which it handles
-		// internally for performing comparisions. We do our best to represent these as an RFC3339
-		// timestamp string here, as the smallest possible timestamp in the case of negative
-		// infinity and the largest possible for infinity. There is also a pgtype.None infinity
-		// modifier which is being left unhandled currently as I don't know that it has any meaning
-		// for a captured value.
-		if x == pgtype.Infinity {
-			return infinityTimestamp, nil
-		} else if x == pgtype.NegativeInfinity {
-			return negativeInfinityTimestamp, nil
-		}
-	case time.Time:
-		return formatRFC3339(x)
-	case pgtype.Time:
-		return x.Microseconds, nil // For historical reasons, times (note: not timestamps) without time zone are serialized as Unix microseconds
-	case pgtype.Numeric:
-		// By default a Numeric value is marshalled to a JSON number, but most JSON parsers and
-		// encoders suffer from precision issues so we really want them to be turned into a JSON
-		// string containing the formatted number instead.
-		var bs, err = json.Marshal(x)
-		if err != nil {
-			return nil, err
-		}
-		return string(bs), nil
-	case pgtype.Bits:
-		return x.Value() // The Value() method encodes to the desired "0101" text string
-	case pgtype.Interval:
-		return x.Value()
-	case pgtype.Line:
-		return x.Value()
-	case pgtype.Lseg:
-		return x.Value()
-	case pgtype.Box:
-		return x.Value()
-	case pgtype.Path:
-		return x.Value()
-	case pgtype.Polygon:
-		return x.Value()
-	case pgtype.Circle:
-		return x.Value()
-	}
-
-	return val, nil
-}
-
-func stringifyRange(r pgtype.Range[any]) (string, error) {
-	if r.LowerType == pgtype.Empty || r.UpperType == pgtype.Empty {
-		return "empty", nil
-	}
-
-	var buf = new(strings.Builder)
-	switch r.LowerType {
-	case pgtype.Inclusive:
-		buf.WriteString("[")
-	case pgtype.Exclusive, pgtype.Unbounded:
-		buf.WriteString("(")
-	}
-	if r.LowerType == pgtype.Inclusive || r.LowerType == pgtype.Exclusive {
-		if translated, err := translateRecordField(nil, r.Lower); err != nil {
-			fmt.Fprintf(buf, "%v", r.Lower)
-		} else {
-			fmt.Fprintf(buf, "%v", translated)
-		}
-	}
-	buf.WriteString(",")
-	if r.UpperType == pgtype.Inclusive || r.UpperType == pgtype.Exclusive {
-		if translated, err := translateRecordField(nil, r.Upper); err != nil {
-			fmt.Fprintf(buf, "%v", r.Upper)
-		} else {
-			fmt.Fprintf(buf, "%v", translated)
-		}
-	}
-	switch r.UpperType {
-	case pgtype.Inclusive:
-		buf.WriteString("]")
-	case pgtype.Exclusive, pgtype.Unbounded:
-		buf.WriteString(")")
-	}
-	return buf.String(), nil
-}
-
-func formatRFC3339(t time.Time) (any, error) {
-	if t.Year() < 0 || t.Year() > 9999 {
-		// We could in theory clamp excessively large years to positive infinity, but this
-		// is of limited usefulness since these are never real dates, they're mostly just
-		// dumb typos like `20221` and so we might as well normalize errors consistently.
-		return negativeInfinityTimestamp, nil
-	}
-	return t.Format(time.RFC3339Nano), nil
-}
-
-func translateArray(_ *sqlcapture.ColumnInfo, x pgtype.Array[any]) (any, error) {
-	var dims []int
-	for _, dim := range x.Dims {
-		dims = append(dims, int(dim.Length))
-	}
-	for idx := range x.Elements {
-		var translated, err = translateRecordField(nil, x.Elements[idx])
-		if err != nil {
-			return nil, err
-		}
-		x.Elements[idx] = translated
-	}
-	return map[string]any{
-		"dimensions": dims,
-		"elements":   x.Elements,
-	}, nil
 }
 
 type columnSchema struct {
